feature!: make experimental encoding deafault
