+++ examples/contracts/Cargo.toml
@@ -17,4 +17,4 @@ tokio = { workspace = true, features = ["full"] }
 [features]
 fuel-core-lib = ["fuels/fuel-core-lib"]
 rocksdb = ["fuels/rocksdb"]
-experimental = ["fuels/experimental"]
+legacy_encoding = ["fuels/legacy_encoding"]
