+++ examples/debugging/src/lib.rs
@@ -1,16 +1,16 @@
 #[cfg(test)]
 mod tests {
-    #[cfg(not(feature = "experimental"))]
+    #[cfg(feature = "legacy_encoding")]
     use std::collections::HashMap;
 
-    #[cfg(not(feature = "experimental"))]
+    #[cfg(feature = "legacy_encoding")]
     use fuel_abi_types::abi::program::ProgramABI;
-    #[cfg(not(feature = "experimental"))]
+    #[cfg(feature = "legacy_encoding")]
     use fuels::core::codec::{calldata, fn_selector};
-    #[cfg(not(feature = "experimental"))]
+    #[cfg(feature = "legacy_encoding")]
     use fuels::types::{errors::Result, param_types::ParamType, SizedAsciiString};
 
-    #[cfg(not(feature = "experimental"))]
+    #[cfg(feature = "legacy_encoding")]
     #[test]
     fn get_a_fn_selector() {
         use fuels::core::{codec::resolve_fn_selector, traits::Parameterize};
@@ -26,7 +26,7 @@ mod tests {
         // ANCHOR_END: example_fn_selector
     }
 
-    #[cfg(not(feature = "experimental"))]
+    #[cfg(feature = "legacy_encoding")]
     #[test]
     fn a_fn_selector_from_json_abi() -> Result<()> {
         use fuels::core::codec::resolve_fn_selector;
@@ -64,7 +64,7 @@ mod tests {
         Ok(())
     }
 
-    #[cfg(not(feature = "experimental"))]
+    #[cfg(feature = "legacy_encoding")]
     #[test]
     fn test_macros() -> Result<()> {
         let function_selector = fn_selector!(initialize_counter(u64));
