+++ packages/fuels-core/Cargo.toml
@@ -37,4 +37,4 @@ tokio = { workspace = true, features = ["test-util", "macros"] }
 [features]
 default = ["std"]
 std = ["dep:fuel-core-client"]
-experimental = []
+legacy_encoding = []
