+++ packages/fuels-core/src/types.rs
@@ -17,9 +17,9 @@ pub mod unresolved_bytes;
 mod wrappers;
 
 pub type ByteArray = [u8; 8];
-#[cfg(not(feature = "experimental"))]
+#[cfg(feature = "legacy_encoding")]
 pub type Selector = ByteArray;
-#[cfg(feature = "experimental")]
+#[cfg(not(feature = "legacy_encoding"))]
 pub type Selector = Vec<u8>;
 
 /// Converts a u16 to a right aligned array of 8 bytes.
