+++ packages/fuels-programs/Cargo.toml
@@ -29,4 +29,4 @@ tempfile = "3.8.1"
 [features]
 default = ["std"]
 std = ["fuels-core/std", "fuels-accounts/std"]
-experimental = ["fuels-core/experimental"]
+legacy_encoding = ["fuels-core/legacy_encoding"]
