+++ packages/fuels/tests/logs.rs
@@ -1453,7 +1453,7 @@ async fn can_configure_decoder_for_script_log_decoding() -> Result<()> {
 // String slices cannot be decoded from logs as they are encoded as ptr, len
 // TODO: Once https://github.com/FuelLabs/sway/issues/5110 is resolved we can remove this
 #[tokio::test]
-#[cfg(not(feature = "experimental"))]
+#[cfg(feature = "legacy_encoding")]
 async fn string_slice_log() -> Result<()> {
     setup_program_test!(
         Wallets("wallet"),
@@ -1486,7 +1486,7 @@ async fn string_slice_log() -> Result<()> {
 }
 
 #[tokio::test]
-#[cfg(feature = "experimental")]
+#[cfg(not(feature = "legacy_encoding"))]
 async fn contract_experimental_log() -> Result<()> {
     use fuels_core::types::AsciiString;
 
@@ -1545,7 +1545,7 @@ async fn contract_experimental_log() -> Result<()> {
 }
 
 #[tokio::test]
-#[cfg(feature = "experimental")]
+#[cfg(not(feature = "legacy_encoding"))]
 async fn script_experimental_log() -> Result<()> {
     use fuels_core::types::AsciiString;
