+++ packages/fuels/tests/providers.rs
@@ -330,9 +330,9 @@ async fn test_gas_forwarded_defaults_to_tx_limit() -> Result<()> {
     );
 
     // The gas used by the script to call a contract and forward remaining gas limit.
-    #[cfg(not(feature = "experimental"))]
+    #[cfg(feature = "legacy_encoding")]
     let gas_used_by_script = 364;
-    #[cfg(feature = "experimental")]
+    #[cfg(not(feature = "legacy_encoding"))]
     let gas_used_by_script = 876;
     let gas_limit = 225_883;
     let response = contract_instance
