+++ packages/fuels/tests/types_scripts.rs
@@ -323,7 +323,7 @@ async fn script_handles_std_string() -> Result<()> {
     Ok(())
 }
 
-#[cfg(feature = "experimental")]
+#[cfg(not(feature = "legacy_encoding"))]
 #[tokio::test]
 async fn nested_heap_types() -> Result<()> {
     setup_program_test!(
